Update StatInfo.cs - fixed variable type to match database entity

+++ NinjaHive.Contract/DTOs/StatInfo.cs
@@ -17,7 +17,7 @@ namespace NinjaHive.Contract.DTOs
         public int Intelligence { get; set; }
         public int Hunger { get; set; }
         public int Stamina { get; set; }
-        public int Resistance { get; set; }
+        public double Resistance { get; set; }
 
     }
 }
